New: Image zoom(preliminary), Jump to index, looping index
You can now zoom in/out of images by clicking on them. This is a pretty basic implementation, but it does work.

The image "display_image_index" is now a text widget that you can edit to jump to the index of your choice.

You can now "loop" through the index by going backwards from the start, or going forwards from the end.