Interface tweaks, Image scaling tweaked, zoom removed
The image is now much larger if you use the app with a large window size, and properly scales up/down.

The downside is using the app with a smaller window sort of squishes the text box.

Zoom was removed because of errors I couldn't resolve. It was poorly implemented anyway, and images are bigger now so it shouldn't be as much an issue.

Minor tweaks to typo correction.