Autocomplete update - Efficiency increase
Autocomplete has been sped up slightly by doing two things.

Limit the amount of returned suggestions. This value is currently set so high that it actually has little to no effect, but it could be tuned in the future.

When updating the suggestions we now only search over the list of previously returned suggestions.