Create: BatchTagEdit & TagEditor
TagEditor is a tool for analyzing and editing tags across many text files. It creates a dictionary of tags and their file/text position so they can be easily addressed later.

BatchTagEdit is the replacement for batch_tag_delete. BTE has an improved and more versatile UI, along with being able to not only delete tags, but edit them.

BatchTagEdit is a more optimized and efficient version of batch_tag_delete.